% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_viz.R
\name{climate_grid}
\alias{climate_grid}
\title{Download and plot essential climate data}
\usage{
climate_grid(print = TRUE)
}
\arguments{
\item{print}{(boolean) Display climate grid ggplot2 chart, defaults to TRUE. Use FALSE to not display chart.}
}
\value{
Invisibly returns a ggplot2 object with climate grid
}
\description{
Plots a 2x2 grid of carbon, temperature, sea ice, and sea level charts.
}
\details{
\code{warming_stripes} invisibly returns a ggplot2 object with 2x2 grid of carbon, temperature, sea ice, and sea level charts from \code{get_carbon}, \code{get_temp}, \code{get_seaice}, and \code{get_sealevel}.
By default the chart is also displayed. Users may further modify the output ggplot2 chart.
}
\examples{
\donttest{
# Draw grid

grid <- climate_grid() }

}
\author{
Hernando Cortina, \email{hch@alum.mit.edu}
}
