% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hockeystick-package.R
\docType{package}
\name{hockeystick-package}
\alias{hockeystick}
\alias{hockeystick-package}
\title{hockeystick: Download and Visualize Essential Global Heating Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides easy access to essential climate change datasets to non-climate experts. Users can download the latest raw data from authoritative sources and view it via pre-defined 'ggplot2' charts. Datasets include atmospheric CO2, methane, emissions, instrumental and proxy temperature records, CMIP6 projections, sea levels, Arctic/Antarctic sea-ice, Hurricanes, Wildfires, and Paleoclimate data. Sources include: NOAA Mauna Loa Laboratory \url{https://gml.noaa.gov/ccgg/trends/data.html}, Global Carbon Project \url{https://www.globalcarbonproject.org/carbonbudget/}, NASA GISTEMP \url{https://data.giss.nasa.gov/gistemp/}, National Snow and Sea Ice Data Center \url{https://nsidc.org/home}, CSIRO \doi{10.1007/s10712-011-9119-1}, NOAA Laboratory for Satellite Altimetry \url{https://www.star.nesdis.noaa.gov/socd/lsa/SeaLevelRise/} and HURDAT Atlantic Hurricane Database \url{https://www.aoml.noaa.gov/hrd/hurdat/Data_Storm.html}, Vostok Paleo carbon dioxide and temperature data: \doi{10.3334/CDIAC/ATG.009}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cortinah.github.io/hockeystick/}
  \item \url{https://github.com/cortinah/hockeystick}
  \item Report bugs at \url{https://github.com/cortinah/hockeystick/issues}
}

}
\author{
\strong{Maintainer}: Hernando Cortina \email{hch@alum.mit.edu} (\href{https://orcid.org/0000-0001-6790-4870}{ORCID})

Authors:
\itemize{
  \item Hernando Cortina \email{hch@alum.mit.edu} (\href{https://orcid.org/0000-0001-6790-4870}{ORCID})
}

}
\keyword{internal}
