% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_climate.R
\name{plot_cmip6}
\alias{plot_cmip6}
\title{Download and plot essential climate data}
\usage{
plot_cmip6(dataset = get_cmip6(), palette = "Spectral", print = TRUE)
}
\arguments{
\item{dataset}{Name of the tibble generated by \code{get_cmip6}}

\item{palette}{(string) Color palette name passed to RColorBrewer. Defaults to
"Spectral" reversed for temperature-type variables.}

\item{print}{(boolean) Display map, defaults to TRUE.}
}
\value{
Invisibly returns a ggplot2 object with the projection map.
}
\description{
Plots gridded CMIP6 climate projections retrieved using \code{get_cmip6()} with ggplot2.
}
\details{
\code{plot_cmip6} returns a pre-defined ggplot2 world map of projected
values. Users may further modify the output chart.
}
