% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cache.R
\name{print.hockeystick_cache_info}
\alias{print.hockeystick_cache_info}
\title{Display data cache info
Shows filenames and cache file sizes}
\usage{
\method{print}{hockeystick_cache_info}(x, ...)
}
\arguments{
\item{x}{filenames}

\item{...}{Additional parameters}
}
\value{
Display data cache info. No return value, called for side effect.
}
\description{
Display data cache info
Shows filenames and cache file sizes
}
