% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hockeystick-package.R
\docType{package}
\name{hockeystick-package}
\alias{hockeystick}
\alias{hockeystick-package}
\title{hockeystick: Download and Visualize Essential Climate Change Data}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

The hockeystick package provides access to essential climate change datasets. The package includes functions to download and visualize atmospheric carbon dioxide.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cortinah/hockeystick}
  \item Report bugs at \url{https://github.com/cortinah/hockeystick/issues}
}

}
\author{
\strong{Maintainer}: Hernando Cortina \email{hch@alum.mit.edu} (\href{https://orcid.org/0000-0001-6790-4870}{ORCID})

}
\keyword{internal}
