% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hockeystick-package.R
\docType{package}
\name{hockeystick-package}
\alias{hockeystick}
\alias{hockeystick-package}
\title{hockeystick: Download and Visualize Essential Climate Change Data}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Provides easy access to essential climate change datasets to non-climate experts. Users can download the latest raw data from authoritative sources and view it via pre-defined 'ggplot2' charts. Datasets include atmospheric CO2, instrumental and proxy temperature records, sea levels, Arctic/Antarctic sea-ice, and Paleoclimate data. Sources include: NOAA Mauna Loa Laboratory <https://www.esrl.noaa.gov/gmd/ccgg/trends/data.html>, NASA GISTEMP <https://data.giss.nasa.gov/gistemp/>, National Snow and Sea Ice Data Center <https://nsidc.org/data/seaice_index/archives>, CSIRO <http://www.cmar.csiro.au/sealevel/sl_data_cmar.html>, NOAA Laboratory for Satellite Altimetry <https://www.star.nesdis.noaa.gov/socd/lsa/SeaLevelRise/>, Vostok Paleo carbon dioxide and temperature data: <https://cdiac.ess-dive.lbl.gov/trends/co2/vostok.html>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cortinah.github.io/hockeystick/}
  \item \url{https://github.com/cortinah/hockeystick}
  \item Report bugs at \url{https://github.com/cortinah/hockeystick/issues}
}

}
\author{
\strong{Maintainer}: Hernando Cortina \email{hch@alum.mit.edu} (\href{https://orcid.org/0000-0001-6790-4870}{ORCID})

}
\keyword{internal}
