% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_climate.R
\name{plot_cmip6_anom}
\alias{plot_cmip6_anom}
\title{Download and plot essential climate data}
\usage{
plot_cmip6_anom(
  dataset = get_cmip6_anom(),
  zero_centered = FALSE,
  print = TRUE
)
}
\arguments{
\item{dataset}{Name of the tibble generated by \code{get_cmip6_anom}}

\item{print}{(boolean) Display map, defaults to TRUE.}
}
\value{
Invisibly returns a ggplot2 object with the change map.
}
\description{
Plots expected climate change (projection minus 1970-2000 baseline) computed by
\code{get_cmip6_anom()} with a diverging color scale centered on zero.
}
\details{
\code{plot_cmip6_anom} returns a pre-defined ggplot2 world map of expected
change in degrees C, with blue for cooling and red for warming. Users may
further modify the output chart.
}
\examples{
\donttest{
plot_cmip6_anom()
}

}
\references{
See \code{\link{get_cmip6}}.
}
\author{
Hernando Cortina, \email{hch@alum.mit.edu}
}
